% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{auto_extdata}
\alias{auto_extdata}
\title{Create delayed assignments for all datasets in a package}
\usage{
auto_extdata(assign.env = parent.frame())
}
\arguments{
\item{assign.env}{The environment in which to create the delayed assignments.}
}
\description{
This function lists all \code{.rds} files in the \code{extdata/} directory
and calls \code{\link{extdata_}} for each.  This offers an easy method to
simulate R's lazy-loading mechanism for data: Place all datasets in
individual \code{.rds} files in \code{extdata/}, and call this function
somewhere in your package's source files.
}
\examples{
\dontrun{
auto_extdata()
}
}

