% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed-assign.R
\name{delayed_assign}
\alias{delayed_assign}
\alias{delayed_assign_}
\title{Create a delayed assignment}
\usage{
delayed_assign(..., assign.env = parent.frame())

delayed_assign_(..., .dots, assign.env = parent.frame())
}
\arguments{
\item{...}{Expressions for which delayed assignments are created.}

\item{assign.env}{The environment in which to create the delayed assignments.}

\item{.dots}{Expressions as lazy objects.}
}
\description{
This function creates one or more delayed assignments in a more natural
way compared to \code{\link{delayedAssign}}.
}
\details{
The "dots" argument can be named or unnamed; for unnamed arguments, names
will be assigned by \code{\link[lazyeval]{all_dots}}.  For each argument,
a delayed assignment with the specified name is created via
\code{\link{delayedAssign}}.  The delayed assignment will evaluate the
argument via \code{\link[lazyeval]{lazy_eval}}.

The function \code{delayed_assign_} is a standard-evaluation version of the
above.  See the "lazyeval" vignette for details.
}
\examples{
\dontrun{
# Creates a delayed assignment "fortytwo" which takes about 7.5 million
# years to compute:
delayed_assign(fortytwo = { Sys.sleep(7.5e6 * 365 * 24 * 60 * 60); 42 })
}
}

