% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{extdata}
\alias{extdata}
\alias{extdata_}
\title{Create a delayed assignment for a dataset}
\usage{
extdata(..., assign.env = parent.frame())

extdata_(..., .dots, assign.env = parent.frame())
}
\arguments{
\item{...}{Base names for RDS files.}

\item{assign.env}{The environment in which to create the delayed assignments.}

\item{.dots}{Base names for RDS files as lazy objects.}
}
\description{
This function creates one or more delayed assignments, focusing on data
stored in the \code{extdata} directory, and therefore especially useful for
package development.  The main advantage over R's internal data loading
mechanism is faster package build time.
}
\details{
The "dots" argument can be named or unnamed; for unnamed arguments, names
will be assigned by \code{\link[lazyeval]{all_dots}}.  The name of the
RDS file is derived from the argument name and passed over to
\code{\link{read_rds}}.
}
\examples{
\dontrun{
# Creates a delayed assignment "fortytwo" which loads the file
# "extdata/fortytwo.rds"
extdata(fortytwo)
}
}

