% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{extdata_}
\alias{extdata_}
\title{Create a delayed assignment for a dataset}
\usage{
extdata_(..., .dots, assign.env = parent.frame())
}
\arguments{
\item{...}{Expressions for which delayed assignments are created.}

\item{.dots}{Expressions as lazy objects.}

\item{assign.env}{The environment in which to create the delayed assignments.}
}
\description{
Create a delayed assignment for a dataset

A version of \code{\link{extdata}} that uses standard
  evaluation and is suitable for programming.
}
\keyword{internal}

