% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-rds.R
\name{read_rds}
\alias{read_rds}
\title{Delay-load an RDS file}
\usage{
read_rds(..., assign.env = parent.frame())
}
\arguments{
\item{...}{Paths to RDS files.}

\item{assign.env}{The environment in which to create the delayed assignments.}
}
\description{
This function creates one or more delayed assignments that load RDS files
(via \code{\link{readRDS}}). For unnamed arguments, names are derived
from the basename of the RDS files.  Paths are normalized (via
\code{\link{normalizePath}}); a warning is thrown if the file
does not exist.
}
\examples{
\dontrun{
# Creates a delayed assignment "fortytwo" which loads the file
# "extdata/fortytwo.rds"
read_rds(fortytwo = system.file("extdata", "fortytwo.rds"))
}
}

