% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rextdata.R
\docType{package}
\name{rextdata-package}
\alias{rextdata}
\alias{rextdata-package}
\title{Simple Access to Data Stored in "extdata" Directory}
\description{
Provides helper functions to access data files in the "extdata"
directory.
}
\author{
Kirill Müller
}

