% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{use_extdata}
\alias{use_extdata}
\title{Use an object as external dataset}
\usage{
use_extdata(..., pkg = ".", compress = "xz", overwrite = FALSE,
  env = parent.frame())
}
\arguments{
\item{...}{objects to save in \code{name} or \code{name = value} format}

\item{compress}{a logical specifying whether saving to a named file is to use
\code{"gzip"} compression, or one of \code{"gzip"}, \code{"bzip2"} or
\code{"xz"} to indicate the type of compression to be used.}

\item{overwrite}{overwrite existing files?}

\item{env}{the environment in which to evaluate the expressions}
}
\description{
Call this function to save an object as \code{.rds} file in the
\code{inst/extdata} directory for later retrieval via \code{\link{extdata}}
or \code{\link{auto_extdata}}.
}

