% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{use_rextdata}
\alias{use_rextdata}
\title{Prepare a package for rextdata}
\usage{
use_rextdata(compress = "xz", pkg = ".", overwrite = FALSE)
}
\arguments{
\item{compress}{a logical specifying whether saving to a named file is to use
\code{"gzip"} compression, or one of \code{"gzip"}, \code{"bzip2"} or
\code{"xz"} to indicate the type of compression to be used.}

\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information.}

\item{overwrite}{overwrite existing files?}
}
\description{
Adds the \code{rextdata} package to the \code{Imports} section of the
\code{DESCRIPTION}, creates a helper file that loads all external data
sets via \code{\link{auto_extdata}}, and converts all existing data sets
to \code{extdata} files.
}

