% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{extdata}
\alias{extdata}
\alias{extdata_}
\title{Create a delayed assignment for a dataset}
\usage{
extdata_(..., .dots, assign.env = parent.frame())
}
\arguments{
\item{...}{Expressions for which delayed assignments are created.}

\item{.dots}{Expressions as lazy objects.}

\item{assign.env}{The environment in which to create the delayed assignments.}
}
\description{
This function creates one or more delayed assignments.  It is focused on data
stored in the \code{extdata} directory, and therefore especially useful for
package development.  The main advantages over R's internal data loading
mechanism are faster package build times; the delayed assignment makes sure
that the data is loaded only if necessary.
}
\details{
The "dots" argument can be named or unnamed; for unnamed arguments, names
will be assigned by \code{\link[lazyeval]{all_dots}}.  For each argument,
a delayed assignment with the specified name is created via
\code{\link{delayedAssign}}.  The delayed assignment will evaluate the
argument; if the argument is a single name, the corresponding \code{.rds}
data file from the \code{extdata} directory will be loaded.

The function \code{\link{extdata_}} is a standard-evaluation version of the
above.  See the "lazyeval" vignette for details.
}
\examples{
\dontrun{
# Creates a delayed assignment "fortytwo" which loads the file
# "extdata/fortytwo.rds"
extdata(fortytwo = readRDS(system.file("extdata", "fortytwo.rds",
        package = "mypackage")))
# A shorter version of the above
extdata(fortytwo)
}
}
\keyword{internal}

