% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{use_extdata_}
\alias{use_extdata_}
\title{Use an object as external dataset}
\usage{
use_extdata_(..., .dots, compress = "xz", pkg = ".", overwrite = FALSE,
  env = parent.frame())
}
\arguments{
\item{...}{Expressions for which delayed assignments are created.}

\item{.dots}{Expressions as lazy objects.}
}
\description{
Use an object as external dataset

A version of \code{\link{extdata}} that uses standard
  evaluation and is suitable for programming.
}
\keyword{internal}

